unify filters and tests and remove logs
